moving all files to separate directory
