+++ libsolidity/boogie/ASTBoogieUtils.cpp
@@ -1,8 +1,8 @@
 #include <boost/algorithm/string/predicate.hpp>
 #include <boost/regex.hpp>
-#include <libsolidity/ast/ASTBoogieUtils.h>
 #include <libsolidity/ast/AST.h>
-#include <libsolidity/ast/BoogieContext.h>
+#include <libsolidity/boogie/ASTBoogieUtils.h>
+#include <libsolidity/boogie/BoogieContext.h>
 #include <libsolidity/ast/Types.h>
 #include <libsolidity/ast/TypeProvider.h>
 #include <liblangutil/SourceLocation.h>
