+++ libsolidity/boogie/BoogieContext.cpp
@@ -1,8 +1,8 @@
 #include <libsolidity/ast/AST.h>
-#include <libsolidity/ast/ASTBoogieUtils.h>
-#include <libsolidity/ast/BoogieContext.h>
-#include <libsolidity/ast/BoogieAst.h>
 #include <libsolidity/ast/TypeProvider.h>
+#include <libsolidity/boogie/ASTBoogieUtils.h>
+#include <libsolidity/boogie/BoogieContext.h>
+#include <libsolidity/boogie/BoogieAst.h>
 
 #include <liblangutil/ErrorReporter.h>
 #include <liblangutil/SourceReferenceFormatter.h>
